LibroFlask/requirements.txt
aniso8601==10.0.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dlib @ file:///C:/Users/Darwin/S2/LibroFlask/dlib-19.22.99-cp310-cp310-win_amd64.whl
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.26.4
opencv-python==4.11.0.86
pillow==11.2.1
PyJWT==2.10.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
